Added delete functionality
The delete feature was implemented on several stages, first implemented a delete_from_json function, in handle_delete file in infrastructure > file_system, then implemented two separate UI elements, one for deletion and one for confirming deletion. The user must provide product name, and after is being asked if they are sure they want to delete the product. The deletion happens so - the program fetches the json file, parses it, deletes the item needed and then overwrites the file with the changes. Additionally, a new screen for the product collection management was implemented.